refact(DomainException): move to shared and rename to BaseException :recycle:
