+++ packages/Sonar/src/modules/users/entities/Cellphone.ts
@@ -1,4 +1,4 @@
-import { Either, left, right } from "@shared/core/either";
+import { Either, left, right } from "@shared/domain/either";
 import { InvalidCellphoneException } from "./exceptions/InvalidCellphoneException";
 
 export class Cellphone {
