--- packages/Sonar/src/modules/users/entities/exceptions/DomainException.ts
@@ -1,3 +1,0 @@
-export interface DomainException {
-  message: string;
-}
