+++ packages/Sonar/src/modules/users/entities/exceptions/InvalidCellphoneException.ts
@@ -1,9 +1,6 @@
-import { DomainException } from "./DomainException";
+import { BaseException } from "../../../../shared/domain/BaseException";
 
-export class InvalidCellphoneException
-  extends Error
-  implements DomainException
-{
+export class InvalidCellphoneException extends Error implements BaseException {
   constructor(cellphone: string) {
     super(`Invalid cellphone ${cellphone}`);
     this.name = "InvalidCellphoneException";
