+++ packages/Sonar/src/modules/users/entities/exceptions/InvalidEmailException.ts
@@ -1,6 +1,6 @@
-import { DomainException } from "./DomainException";
+import { BaseException } from "../../../../shared/domain/BaseException";
 
-export class InvalidEmailException extends Error implements DomainException {
+export class InvalidEmailException extends Error implements BaseException {
   constructor(email: string) {
     super(`Invalid email: ${email}`);
