+++ packages/Sonar/src/modules/users/entities/exceptions/InvalidPasswordException.ts
@@ -1,6 +1,6 @@
-import { DomainException } from "./DomainException";
+import { BaseException } from "../../../../shared/domain/BaseException";
 
-export class InvalidPasswordException extends Error implements DomainException {
+export class InvalidPasswordException extends Error implements BaseException {
   constructor(password: string) {
     super(`Password ${password} is invalid`);
     this.name = "InvalidPasswordException";
