+++ packages/Sonar/src/shared/domain/BaseException.ts
@@ -1,0 +1,3 @@
+export interface BaseException {
+  message: string;
+}
